backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Brennan Beal",email = "brennan@withprotege.ai"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "polars>=1.31.0,<2.0.0",
    "chromadb>=1.0.15,<2.0.0",
    "sentence-transformers>=5.0.0,<6.0.0",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.35.0",
    "python-multipart>=0.0.6"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
